infra/requirements.txt
fastapi
uvicorn
python-dotenv
email-validator
requests